+++ firestore-pdf-generator/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 0.19.0
+
+- Add example values provided in `extension.yaml` files to `PARAMETERS.md` files.
+- Fix http-pdf-generator not accepting content types starting with `application/json` and having a suffix like `; charset=utf-8`. #48
+
 ## Version 0.18.0
 
 - Update dependencies, using latest version of puppeteer.